requirements.txt
numpy==1.22.3
opencv-contrib-python==4.5.5.64
Pillow==9.0.1
